Address PR #1257 review comments (round 2)
- Tighten offset group filter: only drop singleton-coverage groups when
  multiple singletons all originate from the same degenerate equation
  row (preserves legitimate boundary-only lead/lag offsets from distinct rows)
- Move test to tests/integration/kkt/ with integration+slow markers
- Share fixture: single CLI translation reused by both test methods
- Add encoding='utf-8' to read_text() for cross-platform robustness
- Update PR description to match actual heuristic
